ignore disconnected replicas
